In menus.json, data wich are feeding the RS menu
